Refactored storage to fix linting issue, simplified
