--- packages/storage/src/store/BaseKVStore.ts
@@ -1,25 +1,0 @@
-import { KVStore, KVStoreProvider } from "./types";
-
-export class BaseKVStore implements KVStore {
-  constructor(
-    private readonly provider: KVStoreProvider,
-    private readonly _prefix: string
-  ) {}
-
-  async get<T = unknown>(key: string): Promise<T | undefined> {
-    const k = this.prefix() + "/" + key;
-
-    const data = await this.provider.get();
-    return data[k];
-  }
-
-  set<T = unknown>(key: string, data: T | null): Promise<void> {
-    const k = this.prefix() + "/" + key;
-
-    return this.provider.set({ [k]: data });
-  }
-
-  prefix(): string {
-    return this._prefix;
-  }
-}
