+++ packages/storage/src/store/index.ts
@@ -1,5 +1,5 @@
-export * from "./BaseKVStore";
 export * from "./ExtensionKVStore";
 export * from "./IndexedDBKVStore";
+export * from "./LocalKVStore";
 export * from "./Store";
 export * from "./types";
